+++ src/components/ProfileComponents/ProfileProvider/ProfileProvider.jsx
@@ -1,7 +1,7 @@
 import { useSelector } from "react-redux";
 import { useState } from "react";
 import style from './ProfileProvider.module.sass';
-import Form from "../../Form/FormProfileProvider/Form copy"
+import Form from "../../Form/FormProfileProvider/Form"
 import defaultImage from '../../../assets/image/PerfilImage.png';
 
 function ProfileProvider() {
